Partial implementation of NSInvocation copying function.
NSInvocation's actual copy method (or rather, it's copyWithZone:
method) only seems to copy the method signature to the new object;
the copy has none of the method arguments of the original. To avoid
disrupting code that depends on the standard behaviour, I have
chosen not to implement this as a category on the NSInvocation class,
but it would be easy to turn it into one, if desired.

The implementation is not yet complete; the next step is to actually
copy the arguments, however, that requires a way to determine how
large each argument is. The original class provides a method for
getting the size of the return value, but not of the input
arguments, so I need to add one.