Don't attempt to copy return value of NSInvocation for methods with void return types.
I believe that this was the cause of crashes I was seeing, on the
call to *getReturnValue:*, surprisingly. The docs say the behaviour
of getReturnValue: before calling setReturnValue: is undefined, but
it seems to work fine (and there is no way for the copying routine
to tell if the return value has been set in the original or not).
If we do run into problems, we'll have to simply drop copying return
values, which is probably fine as I don't foresee an actual use for
it in this application. I just did it for completeness. But the
crashes seem only to happen with void-return methods, which is
easy to work around.