Exposed the list of invocations a spy has received and moved filtering method to KWMessagePattern class.
I was needing to add more methods to the KWSpy class, to perform
actions on the received invocations list, which did not seem to be
the responsibility of the spy class. A separate InvocationsList
class might be a useful place to keep such functionality, but for
now, I was able to move the filtering responsibility to the
KWMessagePattern class (it just applies itself to every member of a
given list) and left one case that didn’t fit into this ---
extracting the list of message names --- to the matcher in which it
is used.