Added +canMatchSubject: to all "haveReceived..." matchers to filter non-KWSpy subjects.
Based on comments in the code, I may have incorrectly previously
written this as an instance method, and found it to do nothing, so
I removed. None of the included matchers use it, but while looking
at the verifier code, I found that the method was in fact used, so
I've added it back. I'm not 100% sure that all paths to using a
matcher normally will go through this test, so I've left the type
check in the "evaluate" method as well, but it's possible that it
might be unnecessary.