Introduced a weak ref for target of messages saved by a KWSpy, to avoid leaks.
Since the invocation array is owned by the spy, and the elements (NSInvocations) are retained by the array, having an invocation retain the spy (which will be the target of every message it captures) will result in a retain cycle, and the spy and all the invocations will be leaked at the end of the test.

The change here does not completely eliminate the problem, as it only replaces the target. If the spy appears as any other argument, or is part of a chain of strong references from any of the arguments, a retain cycle can still occur. I can think of no reasonable way to scan all arguments to prevent this, and it's also not workable to *not* retain the arguments, as they may be released before the invocation is later tested (although, in practice, most will probably still be in the auto-release pool, so would technically be safe). Ultimately, the invocations should be stored somewhere else to completely solve the issue.