Copy block arguments to be saved in spy messages.
The previous commit revealed an issue in how block arguments were saved. When block arguments were of the stack variety, they would often be invalid by the time the NSInvocation that pointed to them was finally released. This would cause a crash when the block was sent a "release" message itself. (Prior to the previous change, spies and their NSInvocations were never getting released, so the error never showed up.)

We now scan the arguments and make copies of any block arguments, before retaining the copy. This will not help if some argument holds a reference to a block, but any object that keeps a block reference *should* have copied it already, so this should never be a problem in practice.