Create separate target for copying public headers
Both the Kiwi and Kiwi-XCTest targets include a "Copy Files" build phase to copy the same set of headers to the output directory, and keeping the two separate lists in sync is a pain. (Also, it looks like the Kiwi-OSX target should be doing the same copying as well.) The new target contains only the Copy Files phase and the other targets now depend on that target, so there's only one list to maintain.